requirements.txt
playwright
pytest
pytest-playwright
pytest-xdist
pytest-rerunfailures
allure-pytest
pytest-html
python-dotenv
requests
pytest-bdd
websockets
